requirements.txt
click>=8.0
logzero
reorder_editable
pyyaml
